[#382] Separate base logic class
(No functional changes)
